requirements.txt
Django==6.0.4
Pillow>=10.1.0
numpy>=1.24.3
opencv-python>=4.8.0
google-generativeai>=0.8.0
